Add Game type for playing the game
The first player in Hex can force a win, so to keep the game balanced, the second player is usually given the option to switch sides. This is called the "pie rule" (https://en.wikipedia.org/wiki/Pie_rule)

To enforce the pie rule, I'm making all game playing go through a Game type and use a common PlayGameMove method. PlayGameMove errors if we forget the rule.